follow a more microservice approach
