+++ chat-sd/app/package.json
@@ -19,7 +19,6 @@
         "amqplib": "^0.10.3",
         "babel-plugin-module-resolver": "^4.1.0",
         "body-parser": "^1.19.0",
-        "cassandra-driver": "^4.7.2",
         "clear": "^0.1.0",
         "cookie-parser": "^1.4.5",
         "cors": "^2.8.5",
