+++ chat-ws/app/package.json
@@ -20,7 +20,6 @@
         "axios": "^1.5.1",
         "babel-plugin-module-resolver": "^4.1.0",
         "body-parser": "^1.19.0",
-        "cassandra-driver": "^4.7.2",
         "cookie-parser": "^1.4.5",
         "cors": "^2.8.5",
         "dotenv": "^8.2.0",
