+++ messages-service/app/package.json
@@ -34,6 +34,7 @@
         "ws": "^8.14.2"
     },
     "devDependencies": {
+        "@types/amqplib": "^0.10.3",
         "@types/clear": "^0.1.2",
         "@types/cookie-parser": "^1.4.4",
         "@types/cors": "^2.8.14",
