+++ messages-service/app/src/worker.ts
@@ -1,0 +1,36 @@
+import amqp, { ConsumeMessage } from 'amqplib'
+import { WSRes, Message } from '@shared/types'
+import DataModel from './model/dataModel'
+
+const dataModel = new DataModel()
+
+export const startWorker = async () => {
+    const rabbitMQConnection = await amqp.connect('amqp://localhost')
+    const rabbitMQChannel = await rabbitMQConnection.createChannel()
+
+    const userQueue = `messaging-service-worker`
+    await rabbitMQChannel.assertQueue(userQueue, { durable: false })
+
+    rabbitMQChannel.consume(
+        userQueue,
+        async (message: ConsumeMessage | null) => {
+            if (message === null) return
+            const msg = JSON.parse(message.content.toString()) as WSRes
+            const parseBuffer = msg.data as Message
+
+            // Depending on the message type, invoke the corresponding handling function.
+            switch (msg.type) {
+                case 'newMessage':
+                    dataModel.insertMessage(parseBuffer)
+                    break
+                case 'received':
+                    dataModel.setReceived(parseBuffer)
+                    break
+                case 'read':
+                    dataModel.setRead(parseBuffer)
+                    break
+            }
+            rabbitMQChannel.ack(message) // Acknowledge the processed message.
+        }
+    )
+}
