+++ presence-server/app/src/server.ts
@@ -6,6 +6,8 @@ import cors from 'cors'
 import bodyParser from 'body-parser'
 import cookieParser from 'cookie-parser'
 import http from 'http'
+const STATUS_CHECK_INTERVAL = 4000
+import { statusChecker } from './statusChecker'
 
 const app = express()
 const port = process.env.PORT || 7777
@@ -30,6 +32,8 @@ app.get('/health-check', (req, res, next) =>
 app.get('*', (req, res) =>
     res.status(404).send('404. This endpoint does not exist:')
 )
+
+setInterval(statusChecker, STATUS_CHECK_INTERVAL)
 
 // Create an HTTP server
 const server = http.createServer(app)
